chore: Rename cpfp.rs -> keyed_anchor.rs
