+++ crates/connectors/src/prelude.rs
@@ -1,6 +1,6 @@
 //! This module exports all connectors in this crate for convenience.
 
 pub use super::{
-    claim_contest::*, claim_payout::*, contest_counterproof::*, cpfp::*, n_of_n::*, p2a::*,
+    claim_contest::*, claim_payout::*, contest_counterproof::*, keyed_anchor::*, n_of_n::*, p2a::*,
     timelocked::*, unstaking_intent::*,
 };
